Refactored Auth logic to split it in a service and the blueprint to allow decoupling when using flask-restx
