+++ src/services/__init__.py
@@ -1,0 +1,2 @@
+from src.services.auth_service import AuthService
+from src.services.log_service import LogService
